Deprecate two-way toggle and add one-way toggle
This moves the functionality of the two-way toggle, previously the `toggle` component to a component named `toggle-old`. The new one-way toggle takes it's place as `toggle`.

From this point forward, all new uses of toggles should be done using the `toggle` component and uses of `toggle-old` should be migrated to use the one-way toggle.